ref: move components to module
